Add some data creation and return the services

+++ models/incidentUpdate.go
@@ -6,9 +6,9 @@ import (
 
 //IncidentUpdate holds an update for the incident
 type IncidentUpdate struct {
-	ID         int64     `json:"id"`
+	ID         int       `json:"id"`
 	Time       time.Time `json:"time"`
-	IncidentID int64     `json:"incidentId"`
+	IncidentID int       `json:"incidentId"`
 	Status     string    `json:"status"`
 	Message    string    `json:"message"`
 }
